Refactors stuff and moves build outpath at the config level to make it easier for cli
